RNRBC-120: Bring results page in line with prototype
