+++ app/uk/gov/hmrc/residencenilratebandcalculator/views/check_answers.scala.html
@@ -10,7 +10,7 @@
 
     <h1>@messages("check_answers.title")</h1>
 
-    @rnrbHelpers.rows(data)
+    @rnrbHelpers.answer_rows(data)
 
     <a href="@resultsPage.url" class="button">@messages("check_answers.button_text")</a>
 }
