+++ app/uk/gov/hmrc/residencenilratebandcalculator/views/rnrbHelpers/answer_rows.scala.html
@@ -1,0 +1,23 @@
+@import uk.gov.hmrc.residencenilratebandcalculator.models.AnswerRow
+
+@(rows: Seq[AnswerRow], includeChangeLink: Boolean = true)(implicit messages: Messages)
+<ul class="tablular-data">
+    @for(row <- rows) {
+        <li class="tabular-data__entry tabular-data__entry--grouped divider--bottom">
+            <div class="tabular-data__data tabular-data__data--6-12">
+                @row.title
+            </div>
+            <div class="tabular-data__data tabular-data__data--3-12 u-align--tablet-right">
+                @row.data
+            </div>
+            @if(includeChangeLink) {
+                <div class="tabular-data__data tabular-data__data--3-12">
+                    <a href="@row.url">
+                        <span aria-hidden="true">@messages("site.change")</span>
+                        <span class="visually-hidden">@messages("site.hidden-change", row.title)</span>
+                    </a>
+                </div>
+            }
+        </li>
+    }
+</ul>
